recover the dataset file to avoid merge error when fixing issue 92
